+++ FlightBookingApp/src/Components/CardFlight/PlusButton.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { Texto, Plus, } from '../../Assets/styled';
+
+export function PlussButton(){
+  
+    return(
+        <Plus>
+          <Texto 
+            size={'60px'} 
+            color={'white'} 
+            align={'center'} 
+            MP={'-4% 0% 0% 0%'} 
+            >
+            +
+          </Texto>
+        </Plus>
+    );
+}
