+++ FlightBookingApp/src/Components/CardFlight/index.js
@@ -1,0 +1,3 @@
+export { CardFlight } from './CardFlight';
+export {PlussButton} from './PlusButton';
+export {MyFlights_List} from './My_flights_list';
