--- FlightBookingApp/src/Components/Tarjetaprueba.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { Text } from 'react-native'
-
-export const Tarjetaprueba = (props) => {
-    return (
-        <Text>
-            {JSON.stringify(props.info)}
-        </Text>
-    )
-}
